+++ sf_daq_broker/rest_api.py
@@ -1,12 +1,11 @@
-import json
-import logging
-
 import bottle
 
-_logger = logging.getLogger(__name__)
+from .rest_api_error import register_error_handler
 
 
 def register_rest_interface(app, manager):
+
+    register_error_handler(app)
 
     @app.post("/retrieve_from_buffers")
     def retrieve_from_buffers():
@@ -47,15 +46,3 @@ def register_rest_interface(app, manager):
     @app.post("/close_pgroup_writing")
     def close_pgroup_writing():
         return manager.close_pgroup_writing(request=bottle.request.json, remote_ip=bottle.request.remote_addr)
-
-    @app.error(500)
-    def error_handler_500(error):
-        bottle.response.content_type = "application/json"
-        bottle.response.status = 200
-
-        error_text = str(error.exception)
-
-        _logger.error(error_text)
-
-        return json.dumps({"state": "error",
-                           "status": error_text})
