+++ sf_daq_broker/rest_api_slow.py
@@ -1,12 +1,11 @@
-import json
-import logging
-
 import bottle
 
-_logger = logging.getLogger(__name__)
+from .rest_api_error import register_error_handler
 
 
 def register_rest_interface(app, manager):
+
+    register_error_handler(app)
 
     @app.post("/get_detector_settings")
     def get_detector_settings():
@@ -27,15 +26,3 @@ def register_rest_interface(app, manager):
     @app.post("/set_dap_settings")
     def set_dap_settings():
         return manager.set_dap_settings(request=bottle.request.json, remote_ip=bottle.request.remote_addr)
-
-    @app.error(500)
-    def error_handler_500(error):
-        bottle.response.content_type = "application/json"
-        bottle.response.status = 200
-
-        error_text = str(error.exception)
-
-        _logger.error(error_text)
-
-        return json.dumps({"state": "error",
-                           "status": error_text})
